[droidian-apt-config] Initial plumbing to make Droidian switch to Debian unstable
Basing on-top of Debian testing makes a lot of sense when the final goal
is a fixed-point release.

After almost four years of development spanning three Debian releases, it's
clear that there isn't neither the interest nor the workforce to properly
maintain a fixed-point release.

Both the previous fixed-point releases (bullseye, bookworm) are abandoned
to their destiny and aren't used in the wild.

This puts Droidian in a awkward situation where every ~2 years the developers
need to prepare the archive for the new upcoming Debian release, and eventually
ask people to upgrade to that.

This commit is the first step in the preparation to the move to the actual
Debian rolling release stream (unstable). Currently, the actual Droidian repository
will stay at the trixie-level, but the Debian ones will move up.

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>